Implemented Strategy Pattern for Cache Provider
